requirements.txt
pandas
numpy
python-dotenv
arize
